% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simfunctions.R
\name{go_program}
\alias{go_program}
\title{Calling a function that will get the simulation code set up}
\usage{
go_program()
}
\description{
The function sets the proper output directories,
gathers the needed input files, and then proceeds to call a secondary
function that will actually rin the simualtion code.
In order to use the function, all the required variables for the
simulation must have already been initialized.
}
