% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simfunctions.R
\name{progress}
\alias{progress}
\title{Progress monitoring}
\usage{
progress(percent, max = 100)
}
\description{
Monitors the progress of the simulation code by percent complete.
Prints to screen the progress.
}
