% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simfunctions.R
\name{run_sim}
\alias{run_sim}
\title{Function calls the C code and runs the simulation}
\usage{
run_sim(combinations, run_type)
}
\arguments{
\item{combinations}{Number of total combinations of variables}

\item{run_type}{Either "serial or "parallel"}
}
\description{
Calls the simulation code in a nested "for" loop that inputs
all possible combinations of the variables. Also allows for either serial
or parallel code to be run.
}
